cli/show: add extra element to structured output for metadata
Initially this will contain the database uuid and last revision,
and a query_type describing the remaining elements

The changes to emacs/notmuch-tree.el and test/* are to (temporarily)
force these to continue to use the old format.

One non-trivial thing already tested is that the text and mbox formats
do not change in format version 3.